BACK_END/requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.15
django-allauth==0.42.0
django-bootstrap-pagination==1.7.1
django-cors-headers==3.3.0
django-extensions==2.2.9
django-rest-auth==0.9.5
djangorestframework==3.11.0
Faker==4.1.0
idna==2.9
oauthlib==3.1.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
<<<<<<< HEAD
=======
text-unidecode==1.3
>>>>>>> 5a28b2b4f34c785554c5cf3e37c4e90d06609753
urllib3==1.25.9
